backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
web3==6.6.0
bcrypt==4.0.1
PyJWT==2.8.0

